Separate update command into own file cmd/jb/update.go
